tests/container_tests/requirements.txt
requests
pytest
pyhamcrest
